requirements.txt
requests
pydantic
rich
google-genai